requirements.txt
streamlit>=1.28.0
azure-ai-documentintelligence>=1.0.0
azure-core>=1.29.0
pandas>=2.0.0
Pillow>=10.0.0
python-dotenv>=1.0.0
PyMuPDF>=1.23.0